Agregado de Latitud y Longitud

+++ src/main/java/com/proyecto_final/proyecto_final/Model/Ruta.java
@@ -14,7 +14,7 @@ public class Ruta {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idRuta;
+    private Long idRuta;
 
     @Column(nullable = false)
     private String nombre;
